# docs/openapi.yaml
openapi: 3.1.0
info:
  title: Waimai CPS API
  version: 1.0.0
servers:
  - url: https://api.example.com/api/v1
paths:
  /mini/auth/login:
    post:
      summary: 微信 code 登录
      security: []
      requestBody:
        required: true
        content:
          application/json:
            schema: {$ref: "#/components/schemas/LoginRequest"}
      responses:
        "200": {$ref: "#/components/responses/Success"}
...
  /admin/auth/login:
    post: {summary: 租户管理员登录, security: [], responses: {"200": {$ref: "#/components/responses/Success"}, "401": {$ref: "#/components/responses/Error"}}}
...
components:
  securitySchemes:
    miniBearer: {type: http, scheme: bearer, bearerFormat: opaque-random-token}
    adminCookie: {type: apiKey, in: cookie, name: ADMIN_SESSION}
...
  schemas:
    LoginRequest: {type: object, required: [code], properties: {code: {type: string, minLength: 1}, scene: {type: string, description: 24 小时有效且一次性消费的邀请场景参数}}}
    ApiResponse: {type: object, required: [code, message, requestId], properties: {code: {type: string}, message: {type: string}, data: {}, requestId: {type: string, format: uuid}}}
  responses:
    Success: {description: 成功, content: {application/json: {schema: {$ref: "#/components/schemas/ApiResponse"}}}}
    Error: {description: 业务错误, content: {application/json: {schema: {$ref: "#/components/schemas/ApiResponse"}}}}